Fix project structure. Move our code to apps/
Before, we had our own `app.py`, and so we did the
application setup. However, now we are simply providing a
page for a bigger app. Therefore we have moved our file to
the apps directory.

We will have to check how to make our use of styles agree with the
whole page, and how to use assets considering we are no longer at
the project root.