Change project structure. Add simple map component
I think it might be worthwhile to mix structure proposed in
`https://dash.plotly.com/urls`, since that way we'll be able
to import `app` to each component that needs to implement
a callback without passing `app` around as an argument to a
function.

The new structure treats `implang_utils` as a parent module
and `data` and `components` as submodules; this allows us
to use `import` and `from...` from `components` to import
the dataframes from `data`. This way, similar to `app`, we can
have access to the (static) state of the dataframes without
having to pass the dataframe or its slices around